isup: Extend the transaction display with a search ISUP tab
Add a "Search ISUP" button and enable it if any host involved
in the current transaction has SIP capabilities. This way one
can always see if ISUP search is available or not and it is
displayed at the same position (easier to remember/learn where
it is).

Merge the ISUP search result with the existing transaction and
resort it.