message display: Format an ISUP message differently
For SIP there are invididual lines already, for ISUP there is only
the raw JSON string. Start to parse the JSON and display the data
in a table like structure. Put some common fields first and then
the rest. Assume that no HTML tags come from the homer-api and that
no json object has a nesting higher two.

Keep the name as sipMessage as the view code seems to use the name.